+++ mrmustard/lab_dev/circuits/circuits.py
@@ -26,7 +26,7 @@ from pydoc import locate
 from mrmustard import math, settings
 from mrmustard.utils.serialize import save
 from mrmustard.lab_dev.circuit_components import CircuitComponent
-import mrmustard.lab_dev.circuit_components_utils.branch_and_bound as bb
+import mrmustard.lab_dev.circuits.branch_and_bound as bb
 
 __all__ = ["Circuit"]
