+++ tests/test_lab_dev/test_circuits.py
@@ -18,10 +18,10 @@
 
 import pytest
 
-import mrmustard.lab_dev.circuit_components_utils.branch_and_bound as bb
+import mrmustard.lab_dev.circuits.branch_and_bound as bb
 from mrmustard import settings
 from mrmustard.lab_dev.circuit_components import CircuitComponent
-from mrmustard.lab_dev.circuits import Circuit
+from mrmustard.lab_dev.circuits.circuits import Circuit
 from mrmustard.lab_dev.states import Coherent, Number, SqueezedVacuum, Vacuum
 from mrmustard.lab_dev.transformations import Attenuator, BSgate, Dgate, Sgate
 from mrmustard.utils.serialize import load
